+++ DeviceParts/src/org/omnirom/device/RefreshRateTileService.java
@@ -78,10 +78,10 @@ public class RefreshRateTileService extends TileService {
             enabled = RefreshRateSwitch.isCurrentlyEnabled(this);
             RefreshRateSwitch.setPeakRefresh(this, enabled);
             sharedPrefs.edit().putBoolean(DeviceSettings.KEY_REFRESH_RATE, enabled ? false : true).commit();
-            Settings.System.putInt(this.getContentResolver(),
-                    Settings.System.PEAK_REFRESH_RATE, enabled ? 60 : 90);
-            Settings.System.putInt(this.getContentResolver(),
-                    Settings.System.MIN_REFRESH_RATE, enabled ? 60 : 90);
+            Settings.System.putFloat(this.getContentResolver(),
+                    Settings.System.PEAK_REFRESH_RATE, enabled ? 60f : 90f);
+            Settings.System.putFloat(this.getContentResolver(),
+                    Settings.System.MIN_REFRESH_RATE, enabled ? 60f : 90f);
             getQsTile().setIcon(Icon.createWithResource(this,
                     enabled ? R.drawable.ic_refresh_tile_60 : R.drawable.ic_refresh_tile_90));
             getQsTile().setState(enabled ? Tile.STATE_INACTIVE : Tile.STATE_ACTIVE);
